Near-complete rewrite, primarily for performance
It now reliably handles hundreds of connections per second, by using 4 concurrent (Task) listeners per CPU. (This number should be configurable, but it isn't.) On the other hand, it now depends on Nito.AsyncEx, and I'm not sure how to package and distribute that.

Major refactoring for clarity and simplification.

Switched to HttpListener and remove a bunch of hand-rolled TCP/HTTP-handling code. This meant having to re-generate the first line by-hand, and to re-serializing the headers, so it's not quite as certain that the output is identical to the input (but it's very likely).

Added a blank line between the headers and the body in the output; this should make the files more replayable.

Added error handling and reporting (to the user and the requestor) in various places. Note that files will still be written even if they generate an error (in most cases).

Fixed the bug with headers that had colons (":") in the value.

Output files into a subdirectory.

Killed off the "-h" argument - it now binds to all interfaces automatically.

Haven't tested the "-f" argument after all these changes.

The "can't start" error message now reminds you that the port might be in use.