Add a "secret" performance mode
(which is silly since this is open-source)

The perf mode doesn't try to search the body for a batch ID, and just
saves the file with a randomly-generated GUID. It also uses subdirectories
to break up the saved files into smaller sets (for really-large-scale
data sets).

Also added rudimentary argument error checking.